Move buildQuery to release utils
